Add Start-GuestConfigurationPackageRemediation (#176)
* Start implementation.

* Add configurationMode into metaconfig and edit some variable names.

* Add publish policy package

* Run tests on both Windows and Linux

* Remove extra policy

* Style changes

* Style Changes

* Remove positional param

* Update to use prerelease

* Update linux

* updating commands with fixes and cleanup (#180)

Adding Amit chmod +x

* Adding monkey patch (#183)

* updating commands with fixes and cleanup
Adding Amit chmod +x

* adding pwsh & GC monkey patch for linux

* potential Pester issue

* potential Pester issue script change in 5.2

* fix pester issue script change in 5.2

* fix pester issue script change in 5.2

* fix pester issue script change in 5.2 with foreach

* fixing pester 5.2 issue & enabling PS7DscSupport exp feature in azdo

* updating pipeline definition for enabling ps7dscsupport

* updating the right azdo pipeline

* Fixed parameter casing for standard params (#181)

* Fixed parameter casing for standard params

* Update Changelog

* add verbose msg enable experimental feature (#184)

* add verbose msg enable experimental feature

* fix Experimental feature PS7DscSupport

* enabling psdsc exp feat

* Skip this test on linux for now

* Update yaml

* Set up CI with Azure Pipelines

[skip ci]

* Minor changes

* changed name of public pipeline

* Update gitversion

* Find module name and update pipeline name

* remove integration tests

* Put tests back

+++ source/Public/Install-GuestConfigurationAgent.ps1
@@ -4,28 +4,55 @@ function Install-GuestConfigurationAgent
     [OutputType([void])]
     param
     (
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Force
     )
 
     # Unzip Guest Configuration binaries
     $gcBinPath = Get-GuestConfigBinaryPath
     $gcBinRootPath = Get-GuestConfigBinaryRootPath
+    $OsPlatform = Get-OSPlatform
 
-    # Clean the bin folder
-    Remove-Item -Path $gcBinRootPath'\*' -Recurse -Force -ErrorAction SilentlyContinue
-
-    $zippedBinaryPath = Join-Path -Path $(Get-GuestConfigurationModulePath) -ChildPath 'bin'
-    if ($(Get-OSPlatform) -eq 'Windows')
+    if ((-not (Test-Path -Path $gcBinPath)) -or $Force.IsPresent)
     {
-        $zippedBinaryPath = Join-Path -Path $zippedBinaryPath -ChildPath 'DSC_Windows.zip'
+        # Clean the bin folder
+        Write-Verbose -Message "Removing existing installation from '$gcBinRootPath'."
+        Remove-Item -Path $gcBinRootPath'\*' -Recurse -Force -ErrorAction SilentlyContinue
+        $zippedBinaryPath = Join-Path -Path $(Get-GuestConfigurationModulePath) -ChildPath 'bin'
+
+        if ($OsPlatform -eq 'Windows')
+        {
+            $zippedBinaryPath = Join-Path -Path $zippedBinaryPath -ChildPath 'DSC_Windows.zip'
+        }
+        else
+        {
+            # Linux zip package contains an additional DSC folder
+            # Remove DSC folder from binary path to avoid two nested DSC folders.
+            $null = New-Item -ItemType Directory -Force -Path $gcBinPath
+            $gcBinPath = (Get-Item -Path $gcBinPath).Parent.FullName
+            $zippedBinaryPath = Join-Path $zippedBinaryPath 'DSC_Linux.zip'
+        }
+
+        Write-Verbose -Message "Extracting '$zippedBinaryPath' to '$gcBinPath'."
+        [System.IO.Compression.ZipFile]::ExtractToDirectory($zippedBinaryPath, $gcBinPath)
+
+        if ($OsPlatform -ne 'Windows')
+        {
+            # Fix for “LTTng-UST: Error (-17) while registering tracepoint probe. Duplicate registration of tracepoint probes having the same name is not allowed.”
+            Get-ChildItem -Path $gcBinPath -Filter libcoreclrtraceptprovider.so -Recurse | ForEach-Object {
+                Remove-Item $_.FullName -Force -ErrorAction SilentlyContinue
+            }
+
+            Get-ChildItem -Path $gcBinPath -Filter *.sh -Recurse | ForEach-Object -Process {
+                chmod @('+x', $_.FullName)
+            }
+        }
+
+        Install-GuestConfigurationMonkeyPatch
     }
     else
     {
-        # Linux zip package contains an additional DSC folder
-        # Remove DSC folder from binary path to avoid two nested DSC folders.
-        $null = New-Item -ItemType Directory -Force -Path $gcBinPath
-        $gcBinPath = (Get-Item -Path $gcBinPath).Parent.FullName
-        $zippedBinaryPath = Join-Path $zippedBinaryPath 'DSC_Linux.zip'
+        Write-Verbose -Message "Guest Configuration Agent binaries already installed at '$gcBinPath', skipping."
     }
-
-    [System.IO.Compression.ZipFile]::ExtractToDirectory($zippedBinaryPath, $gcBinPath)
 }
